agent_src/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-genai
pydantic
